Moved homepage to a home folder in the pages folder
